+++ app/windows/windowsdarktitlebar.h
@@ -1,0 +1,10 @@
+#ifndef WINDOWSDARKTITLEBAR_H
+#define WINDOWSDARKTITLEBAR_H
+
+#include <QWidget>
+
+namespace WindowsDarkTitleBar {
+  void apply(QWidget *window);
+}
+
+#endif // WINDOWSDARKTITLEBAR_H